refactor(physics): flatten LinearSpring/LinearDamper into their ECS components
Per PR review feedback: a standalone LinearSpring/LinearDamper class
wrapped by its ECS component, with the actual force math abstracted
into a separate applyLinearSpringForce/applyLinearDamperForce
function, added indirection the systems didn't need - that's the
whole point of a system. Removes src/physics/forces/ entirely and
inlines bodyA/bodyB/anchorA/anchorB/stiffness (or dampingCoefficient)
directly into LinearSpringEcsComponent/LinearDamperEcsComponent, with
restLength/anchor defaulting moved into the add*Component factories
and the force computation moved directly into
createLinearSpringEcsSystem/createLinearDamperEcsSystem, matching how
AngularVelocityMotorEcsComponent/System already work in this codebase.
