--- src/physics/forces/index.ts
@@ -1,4 +1,0 @@
-export * from './apply-linear-damper-force.js';
-export * from './apply-linear-spring-force.js';
-export * from './linear-damper.js';
-export * from './linear-spring.js';
