+++ src/physics/systems/linear-spring.system.ts
@@ -4,12 +4,16 @@ import {
   LinearSpringEcsComponent,
   LinearSpringId,
 } from '../components/index.js';
-import { applyLinearSpringForce } from '../forces/index.js';
 
 /**
- * Creates an ECS system that applies each matched entity's `LinearSpring`
- * restoring force to its connected bodies every tick, via
- * `applyLinearSpringForce`.
+ * Creates an ECS system that applies each matched entity's
+ * `LinearSpringEcsComponent` restoring force to `bodyA`/`bodyB` every tick:
+ * `F = -k * x`, where `x` is the signed difference between the current
+ * distance separating the two anchors and `restLength`, acting along the
+ * line between them. The force is scaled by `deltaTimeInSeconds` and
+ * applied via `RigidBody.applyImpulse` (the same continuous-force-via-
+ * scaled-impulse pattern used for gravity and wind, see the Applying Forces
+ * guide).
  *
  * Must be registered with the `EcsWorld` before `createPhysicsSyncEcsSystem`
  * (or with an earlier `registrationOrder`), so the force applied this tick
@@ -22,8 +26,26 @@ export const createLinearSpringEcsSystem = (
 ): EcsSystem<[LinearSpringEcsComponent], void> => ({
   query: [LinearSpringId],
   run: (result) => {
-    const [springComponent] = result.components;
+    const [spring] = result.components;
+    const { bodyA, bodyB, anchorA, anchorB, restLength, stiffness } = spring;
+    const { deltaTimeInSeconds } = time;
 
-    applyLinearSpringForce(springComponent.spring, time.deltaTimeInSeconds);
+    const rA = anchorA.rotate(bodyA.angle);
+    const rB = anchorB.rotate(bodyB.angle);
+    const delta = bodyB.position.add(rB).subtract(bodyA.position.add(rA));
+    const length = delta.magnitude();
+
+    if (length === 0) {
+      return;
+    }
+
+    const direction = delta.multiply(1 / length);
+    const displacement = length - restLength;
+    const impulse = direction.multiply(
+      -stiffness * displacement * deltaTimeInSeconds,
+    );
+
+    bodyA.applyImpulse(impulse.multiply(-1), rA);
+    bodyB.applyImpulse(impulse, rB);
   },
 });
